Just updating Y position on Source Rectangle if sprite is animated
There was a bug where on a sprite that was not animated, its Y position would explode to a really high number.

This was happening because m_CurrentAnimationIndex would never be initialized (sprite was not animated) and would contain some trash value (in this specific case, 205) - and this value would multiply the sprite size and pass this value to the Y position of the source rectangle of the sprite.

This was fixed by simply only changing the source rectangle if the sprite is animated.